pyproject.toml
[tool.poetry]
name = "mysite"
version = "1.0.0"
description = "A peculiar Django app"
authors = ["Samuel Herrero <samuel.herrero@feedzai.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
elasticsearch-dsl = "^7.4.1"
django-elasticsearch-dsl = "^7.0"
psycopg2-binary = "^2.9.9"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
